% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_lipidomeR_dependencies.R
\name{install_lipidomeR_dependencies}
\alias{install_lipidomeR_dependencies}
\title{Install dependencies for the lipidomeR.}
\usage{
install_lipidomeR_dependencies(pkgs.bioconductor = c("limma"), ...)
}
\arguments{
\item{pkgs.bioconductor}{Character vector of Bioconductor packages to install.}

\item{...}{Other arguments to the \code{BiocManager::install()} function.}
}
\description{
Install dependencies for the lipidomeR.
}
