% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_lipidomeR_dependencies.R
\name{install_Bioconductor_dependencies}
\alias{install_Bioconductor_dependencies}
\title{Install the Bioconductor dependencies for the 'lipidomeR'}
\usage{
install_Bioconductor_dependencies(pkgs.bioconductor = c("limma"))
}
\arguments{
\item{pkgs.bioconductor}{Character vector of Bioconductor packages to
install.}
}
\description{
Run this function, if you encounter an error with missing packages,
   such as, \href{https://doi.org/doi:10.18129/B9.bioc.limma}{limma}.
}
